feat(kuma-dp): detect memory only on linux
