+++ app/kuma-dp/pkg/dataplane/envoy/memory_limit_darwin.go
@@ -1,0 +1,5 @@
+package envoy
+
+func DetectMaxMemory() uint64 {
+	return 0
+}
